Initial commit
Using vite and react, then @react-three/fiber as the main rendering
engine, jscad for the backend modeling. This commit brings us as far as
being able to define some geometry then pass it to the ThreeRenderer
component which then spits out a full-screen canvas.

There are a couple of additional libraries here yet unused, for one leva,
which I intend to lean on for adding controls to the application.